+++ python_filmaffinity/config.py
@@ -1,0 +1,2 @@
+FIELDS_MOVIE = ['title', 'id']
+FIELDS_TYPE = ['title', 'director', 'cast']