+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+v0.0.4 (12-06-2017)
+*******************
+
+- Top new DVDs
+- Get movie with trailer
+- Top TV series
+- Return movies list with raiting
+
 
 v0.0.3 (10-06-2017)
 *******************